backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0
psycopg2-binary==2.9.9
faker==19.12.0
# openai==1.3.0  # Optional - only needed for AI message generation
